Website Codes/Website/flask/requirements.txt
Flask
Flask-Mail
qrcode
Pillow
python-dotenv